Finalized migration with switch to DataSketches from Apache
