+++ doc/3rd-party-licenses/datasketches-java/NOTICE
@@ -1,0 +1,12 @@
+Apache DataSketches Java
+Copyright 2024 The Apache Software Foundation
+
+Copyright 2015-2018 Yahoo Inc.
+Copyright 2019-2020 Verizon Media
+Copyright 2021 Yahoo Inc.
+
+This product includes software developed at
+The Apache Software Foundation (http://www.apache.org/).
+
+Prior to moving to ASF, the software for this project was developed at
+Yahoo Inc. (https://developer.yahoo.com).
