--- doc/3rd-party-licenses/hash4j/NOTICE
@@ -1,5 +1,0 @@
-hash4j (Dynatrace hash library for Java)
-https://github.com/dynatrace-oss/hash4j
-Copyright Dynatrace LLC
-
-Apache License 2.0
