+++ pom.xml
@@ -544,9 +544,9 @@
         </dependency>
         <!-- https://github.com/dynatrace-oss/hash4j -->
         <dependency>
-            <groupId>com.dynatrace.hash4j</groupId>
-            <artifactId>hash4j</artifactId>
-            <version>0.29.0</version>
+            <groupId>org.apache.datasketches</groupId>
+            <artifactId>datasketches-java</artifactId>
+            <version>6.0.0</version>
         </dependency>
         <!-- Java EE libs -->
         <dependency>
